+++ bundle/Action/CrucialActionInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Netgen\Bundle\InformationCollectionBundle\Action;
+
+interface CrucialActionInterface
+{
+
+}