+++ bundle/Priority.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Netgen\Bundle\InformationCollectionBundle;
+
+class Priority
+{
+    const DEFAULT_PRIORITY = 100;
+}